jpa relations

+++ src/main/java/com/company/foxtask/model/repository/RoomRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface RoomRepository extends JpaRepository<Room, Long> {
+public interface RoomRepository extends JpaRepository<Room, Integer> {
+
     List<Room> findAllByCategory(String category);
 }
